fix: 중복닉네임 처리 개선 (#57)
